+++ libraries/datetime/mod.ts
@@ -1,1 +1,2 @@
-export {};
+export { Timestamp } from './src/Timestamp.ts';
+export { TimeUnit } from './src/TimeUnit.ts';