+++ libraries/datetime/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @williamthorsen/toolbelt.datetime
 
+## 0.2.0
+
+### Features
+
+Added classes:
+
+- `Timestamp`: Represents a point in time.
+- `TimeUnit`: Represents a unit of time and can convert to other time units.
+
 ## 0.1.0
 
 ### Tooling